pyproject.toml
[tool.poetry]
name = "sql_alchemy"
version = "0.1.0"
description = ""
authors = ["Logan Jones <loganemilyjones@mac.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.5"
psycopg2-binary = "^2.8.6"
fire = "^0.4.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
